backend/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
matplotlib
streamlit
plotly
nltk
textblob
sentence-transformers
faiss-cpu
langchain
langchain-community
pypdf
